backend/requirements.txt
fastapi[standard]
uvicorn
python-multipart
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv

